+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Changed
+- Convert functional test fixtures to CSV (#601)
 - Use typed properties instead of `@var` annotations (#599)
 - Return `ResponseInterface` in controller actions (#597)
 - Replace switchable controller actions with separate plugins (#575)